MOVE: rename payload dir to dto
