+++ API/src/main/java/com/litcode/server/controller/userController.java
@@ -1,7 +1,7 @@
 package com.litcode.server.controller;
 
-import com.litcode.server.payload.SignInRequest;
-import com.litcode.server.payload.UserProfile;
+import com.litcode.server.dto.SignInRequest;
+import com.litcode.server.dto.UserProfile;
 import com.litcode.server.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
