+++ API/src/main/java/com/litcode/server/service/UserService.java
@@ -1,7 +1,7 @@
 package com.litcode.server.service;
 
-import com.litcode.server.payload.SignInRequest;
-import com.litcode.server.payload.UserProfile;
+import com.litcode.server.dto.SignInRequest;
+import com.litcode.server.dto.UserProfile;
 
 public interface UserService {
 	UserProfile getUserInfo(SignInRequest signInInfo);
